MC-33831: Issue with oversell of products
