+++ InventoryReservationCli/Test/Integration/_files/order_with_reservation.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
 
 use Magento\Sales\Model\Order\ShipmentFactory;
 
-require __DIR__ . '/../../../../../../../dev/tests/integration/testsuite/Magento/Sales/_files/order.php';
+require '/var/www/magento2/dev/tests/integration/testsuite/Magento/Sales/_files/order.php';
 
 $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
